### 2020-03-09 #2151 @whedon
[PRE REVIEW]: Pydiogment: A Python package for audio augmentation
**Submitting author:** @SuperKogito (<a href="http://orcid.org/0000-0002-5610-9086">Ayoub Malek</a>)
**Repository:** <a href="https://github.com/SuperKogito/pydiogment" target ="_blank">https://github.com/SuperKogito/pydiogment</a>
**Version:** v0.1.0
