### 2020-03-09 #2151 @danielskatz
I've suggested a bunch of small changes in https://github.com/SuperKogito/pydiogment/pull/6, and pointed out one phrase that I don't understand
