### 2020-03-11 #2151 @faroit
@mbobra I would also like to bring up @justinsalamon the author of [scaper](https://github.com/justinsalamon/scaper) and @bmcfee the author of [muda](https://github.com/bmcfee/muda) which are both software packages with a very similar use case. I think it would be important to include one of them as reviewers to reduce the overlap between these packages.
