### 2020-03-18 #2167 @mbobra
@bmcfee @justinsalamon Thank you for agreeing to review this submission! Whedon generated a checklist and linked a reviewer guide above -- please let me know if you have any questions or comments. Feel free to message each other and the submitting author. And I know it is a difficult time to get work done; JOSS' position at the moment is that there is no pressure on reviewers or editors for the next 30 days. So feel free to take your time :)
