### 2020-05-28 SuperKogito/pydiogment#9 @justinsalamon
[JOSS review] installation instructions for macOS and Windows
The installation instructions in the README and the [docs](https://pydiogment.readthedocs.io/en/latest/dependencies.html#installation) assume the user is using a linux OS that supports `apt install`, which isn't the case for windows and macOS.

Please add instructions for installing non-python dependencies on macOS and windows, unless this package only supports linux, in which case it should be clearly stated in the READM/docs.

I have an example of instructions for installing ffmpeg across all 3 OSs here:
https://github.com/justinsalamon/scaper/#non-python-dependencies
