### 2020-05-28 SuperKogito/pydiogment#10 @justinsalamon
[JOSS review] supported audio formats
The README, docs, and docstrings don't specify which audio formats are supported. Please add information to all three about which audio formats are supported.
