### 2020-05-28 SuperKogito/pydiogment#12 @justinsalamon
[JOSS review] fade_in_and_out doesn't work as expected
Just tried `fade_in_and_out()` and got unexpected behavior.

Here are my input/output files (the input is called input.wav): [fade_bug.zip](https://github.com/SuperKogito/pydiogment/files/4698792/fade_bug.zip)

The input is a 2-second, 16-bit, mono, WAV file. The output is heavily distorted and doesn't include the expected fade-in and fade-out behavior.

While on this function, also note it'd be convenient for the user to be able to choose the destination folder for the output as well the filename, or at least a custom suffix as opposed to the hard coded output path and filename.
