### 2020-05-28 SuperKogito/pydiogment#13 @justinsalamon
[JOSS review] gain doesn't work as expected
code:
```python
test_file = "liz.wav"
apply_gain(test_file, +20)
apply_gain(test_file, -20)
```

result: [gain_bug.zip](https://github.com/SuperKogito/pydiogment/files/4698798/gain_bug.zip)
