### 2020-05-28 #2167 @justinsalamon
@arfon @SuperKogito @bmcfee @mbobra

I've started my review (apologies for the delay), but pretty quickly hit some functionality issues.

Here are the issues I've created so far on the target repo:
* Missing installation instructions: https://github.com/SuperKogito/pydiogment/issues/9
* Missing docs on audio formats: https://github.com/SuperKogito/pydiogment/issues/10
* fade issue: https://github.com/SuperKogito/pydiogment/issues/12
* gain issue: https://github.com/SuperKogito/pydiogment/issues/13

Given the last two issues with functionality, should I continue with the review, or wait until they are addressed before continuing? I suspect this issue might impact other parts of the software and would like to avoid duplicating review work if at all possible. Thanks!
