### 2020-06-05 SuperKogito/pydiogment#14 @bmcfee
[JOSS Review] change_tone?
:wave: JOSS reviewer here, following on the main review issue https://github.com/openjournals/joss-reviews/issues/2167

In working through the documentation, I was confused by the `change_tone` function.  I gather that this is a pitch shifting operation, but it's not stated anywhere in the documentation / docstring https://pydiogment.readthedocs.io/en/latest/augf.html#pydiogment.augf.change_tone
