### 2020-06-05 SuperKogito/pydiogment#16 @bmcfee
[JOSS Review] Community guidelines
[review issue](https://github.com/openjournals/joss-reviews/issues/2167)

There are guidelines for contributors in the readme (great!), but there need to also be guidelines for users seeking to report bugs and get help.
