### 2020-06-05 SuperKogito/pydiogment#17 @bmcfee
[JOSS Review] Paper details, related work, experiments
[review issue](https://github.com/openjournals/joss-reviews/issues/2167)

I read through the paper, and have a few comments for improving it.

## Related work

A few projects that aren't, but should be cited:

- Mauch, Matthias, and Sebastian Ewert. "The audio degradation toolbox and its application to robustness evaluation." (2013).  (And python port: https://github.com/sevagh/audio-degradation-toolbox )
- Schlüter, Jan, and Thomas Grill. "Exploring Data Augmentation for Improved Singing Voice Detection with Neural Networks." ISMIR. 2015.

More generally, it's not clear in the writeup how this project compares to the existing alternatives, functionality-wise.  (Full disclosure, there's a bit of an awkward situation here as I'm the author of one of these alternative toolboxes, but I'll try to be objective :grin:.)  I know these papers are meant to be brief, but it's also important to properly establish context, and make clear to readers how this package differs from others.

(As an aside: I don't think the characterization of muda is entirely accurate: we use it for all sorts of things outside of music, notably environmental sound and bioacoustics.)

## Experiment details

It's not clear in the writeup whether your experiment includes augmentation during testing, or only during training.
