### 2020-06-09 #2167 @SuperKogito
@arfon @mbobra  @justinsalamon  @bmcfee,  thank you all for your assistance and reviews :heart:
@justinsalamon  @bmcfee raised quite a few interesting points, that will surely improve the library.
@mbobra I think I can handle most of the required changes. The ones that concern me the most are the functionality issues raised by @justinsalamon. I am not sure yet about the complexity of those, although it is great to know that you got my back. So hopefully this weekend, I will find out if I can handle them and if I need assistance, I will be coming to you; at least for some advice.
@HMMalek, I think we got a lot to do this weekend :smile:
