### 2020-07-14 303a8b2
README update:
Dependencies, bug reports and supported files details.
