### 2020-07-19 ae35497
Create CODE_OF_CONDUCT.md
