### 2020-07-19 1538ac9
Required changes for Code_Of_Conduct.md
add-code-of-conduct-1 PR  changes.
