### 2020-10-26 #2167 @SuperKogito
This is very fair of you, I apologize for the long inactivity. I discussed this with the other contributor and we really like to try to close this -at the latest- by the end of November. We already implemented some of the requested changes, but didn't mention it in the issues yet. So my request is to give us a couple of more weeks as we work on the code and hopefully by November, this will be done & ready either for rejection or acceptance. However, if you think that a resubmission at a later point is a better option at this stage, then I will trust your decision.
