### 2020-11-30 #2167 @SuperKogito
Unfortunately, we didn't manage to implement the requested changes at this point in time. Therefore, I think this should be closed as @arfon suggested and maybe we attempt a resubmission at a later point in time. I would like to thank this community, the editor @mbobra, the reviewers @bmcfee, @justinsalamon for their help, patience and the learning opportunity they provided us during this submission procedure. Maybe this one wasn't fruitful but hopefully the next ones will be. Thank you for your time and energy. Stay safe, stay healthy <3
