### 2020-11-30 #2167 @mbobra
No worries @SuperKogito and thank you for submitting to JOSS! We look forward to your new submission.

@openjournals/joss-eics Can we close (reject) this submission? The authors would like to continue working on it and re-submit at a later date.
